requirements.txt
Django==5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
requests==2.32.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
python-dotenv==1.0.1
django-guardian==2.4.0
django-import-export==4.3.7
pillow==11.2.1
openpyxl==3.1.5
django-phonenumber-field[phonenumberslite]
django-admin-searchable-dropdown
